requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
Django==4.1.1
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-mysql==3.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.0.1
Pillow==8.4.0
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.4
tzlocal==4.2
